+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/StorageSettingsPage.kt
@@ -5,16 +5,25 @@ import android.os.StatFs
 import android.widget.Toast
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.selection.toggleable
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.Card
+import androidx.compose.material.LinearProgressIndicator
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.OutlinedTextField
+import androidx.compose.material.TextButton
 import androidx.compose.material.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -23,6 +32,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
@@ -35,6 +45,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.DialogProperties
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.CallOnce
 import au.com.shiftyjelly.pocketcasts.compose.bars.ThemedTopAppBar
@@ -45,10 +57,14 @@ import au.com.shiftyjelly.pocketcasts.compose.components.SettingRadioDialogRow
 import au.com.shiftyjelly.pocketcasts.compose.components.SettingRow
 import au.com.shiftyjelly.pocketcasts.compose.components.SettingRowToggle
 import au.com.shiftyjelly.pocketcasts.compose.components.SettingSection
+import au.com.shiftyjelly.pocketcasts.compose.components.TextH30
+import au.com.shiftyjelly.pocketcasts.compose.components.TextH40
+import au.com.shiftyjelly.pocketcasts.compose.components.TextH50
 import au.com.shiftyjelly.pocketcasts.compose.components.TextP40
 import au.com.shiftyjelly.pocketcasts.compose.components.TextP50
 import au.com.shiftyjelly.pocketcasts.compose.preview.ThemePreviewParameterProvider
 import au.com.shiftyjelly.pocketcasts.compose.theme
+import au.com.shiftyjelly.pocketcasts.localization.R
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.settings.viewmodel.StorageSettingsViewModel
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
@@ -73,6 +89,7 @@ fun StorageSettingsPage(
         onBackPressed = onBackPressed,
         onClearDownloadCacheClick = { viewModel.onClearDownloadCacheClick() },
         onManageDownloadedFilesClick = onManageDownloadedFilesClick,
+        onFixDownloadsClick = { viewModel.fixDownloadedFiles() },
         bottomInset = bottomInset,
         modifier = modifier,
     )
@@ -117,6 +134,13 @@ fun StorageSettingsPage(
                 Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
             }
     }
+    if (state.fixDownloadsState.isFixing) {
+        FixingDownloadsDialog(
+            state.fixDownloadsState.currentlyFixed,
+            state.fixDownloadsState.episodeCount,
+            onDismiss = { viewModel.cancelDownloadsFix() },
+        )
+    }
 }
 
 @Composable
@@ -125,6 +149,7 @@ fun StorageSettingsView(
     onBackPressed: () -> Unit,
     onClearDownloadCacheClick: () -> Unit,
     onManageDownloadedFilesClick: () -> Unit,
+    onFixDownloadsClick: () -> Unit,
     bottomInset: Dp,
     modifier: Modifier = Modifier,
 ) {
@@ -147,6 +172,7 @@ fun StorageSettingsView(
                         state = state.downloadedFilesState,
                         onClick = onManageDownloadedFilesClick,
                     )
+                    FixDownloads(onFixDownloadsClick)
                     ClearDownloadCacheRow(onClearDownloadCacheClick)
                     StorageChoiceRow(state.storageChoiceState)
                     StorageFolderRow(state.storageFolderState)
@@ -187,6 +213,20 @@ private fun ClearDownloadCacheRow(
 ) {
     SettingRow(
         primaryText = stringResource(LR.string.settings_storage_clear_download_cache),
+        modifier = modifier
+            .clickable { onClick() }
+            .padding(vertical = 6.dp),
+    )
+}
+
+@Composable
+private fun FixDownloads(
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    SettingRow(
+        primaryText = stringResource(R.string.settings_storage_fix_downloads),
+        secondaryText = stringResource(R.string.settings_storage_fix_downloads_description),
         modifier = modifier
             .clickable { onClick() }
             .padding(vertical = 6.dp),
@@ -374,6 +414,63 @@ private fun AlertDialogView(
     )
 }
 
+@Composable
+private fun FixingDownloadsDialog(
+    currentlyFixed: Int,
+    episodeCount: Int,
+    onDismiss: () -> Unit,
+) {
+    Dialog(
+        properties = DialogProperties(
+            dismissOnClickOutside = false,
+        ),
+        onDismissRequest = { onDismiss() },
+    ) {
+        Card(
+            modifier = Modifier
+                .wrapContentWidth()
+                .wrapContentHeight(),
+            shape = RoundedCornerShape(16.dp),
+        ) {
+            Column(
+                modifier = Modifier
+                    .wrapContentHeight()
+                    .padding(vertical = 8.dp),
+                verticalArrangement = Arrangement.Center,
+                horizontalAlignment = Alignment.Start,
+            ) {
+                TextH30(
+                    text = stringResource(R.string.settings_storage_fixing_downloads),
+                    modifier = Modifier.padding(16.dp),
+                )
+                TextH50(
+                    text = stringResource(R.string.settings_storage_fixing_progress, currentlyFixed, episodeCount),
+                    modifier = Modifier.padding(16.dp),
+                )
+                LinearProgressIndicator(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(16.dp),
+                )
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.End,
+                ) {
+                    TextButton(
+                        onClick = { onDismiss() },
+                        modifier = Modifier.padding(end = 8.dp),
+                    ) {
+                        TextH40(
+                            text = stringResource(R.string.cancel).uppercase(),
+                            color = MaterialTheme.theme.colors.primaryInteractive01,
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
+
 private fun mapToStringWithStorageSpace(
     option: String,
     path: String?,
@@ -418,10 +515,14 @@ private fun StorageSettingsPreview(
                 storageDataWarningState = StorageSettingsViewModel.State.StorageDataWarningState(
                     onCheckedChange = {},
                 ),
+                fixDownloadsState = StorageSettingsViewModel.State.FixDownloadsState(
+                    isFixing = false,
+                ),
             ),
             onBackPressed = {},
             onClearDownloadCacheClick = {},
             onManageDownloadedFilesClick = {},
+            onFixDownloadsClick = {},
             bottomInset = 0.dp,
         )
     }