+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/EpisodeManager.kt
@@ -151,4 +151,6 @@ interface EpisodeManager {
 
     suspend fun selectChapterIndexForEpisode(chapterIndex: Int, episode: PodcastEpisode)
     suspend fun deselectChapterIndexForEpisode(chapterIndex: Int, episode: PodcastEpisode)
+
+    suspend fun getAllPodcastEpisodes(limit: Int, offset: Int): List<PodcastEpisode>
 }